requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

# project packages
torch
torchvision
pytest
isort
pre-commit
black
wandb
